Refactor CFX_FontMgr and CFX_FontMapper to break circular dependency
Move font cache maps (face_map_, ttc_face_map_) and FontCacheEntry
from CFX_FontMgr to CFX_FontMapper, as they are exclusively used by
the mapper.

Create core/fxge/cfx_standardfonts.h/.cpp to hold standard font enums,
constants, and raw data helper functions. This removes the need for
CFX_FontMapper to depend on CFX_FontMgr for standard font data.

Use type aliases in CFX_FontMapper to maintain backward compatibility
with other parts of the codebase that reference
CFX_FontMapper::StandardFont.

Update cpdf_stockfontarray.cpp to cast the scoped StandardFont enum
to size_t when indexing.

TAG=agy
CONV=63532648-374f-4a87-acdf-e5aebbaa70a3

+++ core/fpdfapi/font/cpdf_stockfontarray.cpp
@@ -29,13 +29,13 @@ CPDF_StockFontArray::~CPDF_StockFontArray() {
 
 RetainPtr<CPDF_Font> CPDF_StockFontArray::GetFont(
     CFX_FontMapper::StandardFont index) const {
-  CHECK_LT(index, std::size(stock_fonts_));
-  return stock_fonts_[index];
+  CHECK_LT(static_cast<size_t>(index), std::size(stock_fonts_));
+  return stock_fonts_[static_cast<size_t>(index)];
 }
 
 void CPDF_StockFontArray::SetFont(CFX_FontMapper::StandardFont index,
                                   RetainPtr<CPDF_Font> font) {
-  if (index < std::size(stock_fonts_)) {
-    stock_fonts_[index] = std::move(font);
+  if (static_cast<size_t>(index) < std::size(stock_fonts_)) {
+    stock_fonts_[static_cast<size_t>(index)] = std::move(font);
   }
 }
